Aaron 2/24/2023 6:14 PM: Worked on data script and Dev notebook

+++ make_data_structure.sh
@@ -1,14 +1,36 @@
 #!/bin/bash
 
 mkdir -p Data/NHTS_2017
+mkdir -p Data/Generated_Data
 mkdir -p Data/ACS_2021/Tract_Geometries
 
-[ ! -f ACS2021_Table_Shells.csv ] || cp ACS2021_Table_Shells.csv Data/ACS_2021/ACS2021_Table_Shells.csv
+[ ! -f ACS2021_Table_Shells.csv ] ||
+cp ACS2021_Table_Shells.csv Data/ACS_2021/ACS2021_Table_Shells.csv
 
-curl -o Data/ACS_2021/Tract_Geometries/tracts.zip https://www2.census.gov/geo/tiger/GENZ2021/shp/cb_2021_us_bg_500k.zip
+if [ ! -f Data/ACS_2021/Tract_Geometries/tracts.zip ]; then
+	curl -o Data/ACS_2021/Tract_Geometries/tracts.zip https://www2.census.gov/geo/tiger/GENZ2021/shp/cb_2021_us_tract_500k.zip
+	echo "Tracts Downloaded"
+else
+	echo "Tracts Downloaded"
+fi
 
-unzip Data/ACS_2021/Tract_Geometries/tracts.zip -d Data/ACS_2021/Tract_Geometries
+if [ ! -f Data/ACS_2021/Tract_Geometries/cb_2021_us_bg_500k.shp ]; then
+	unzip Data/ACS_2021/Tract_Geometries/tracts.zip -d Data/ACS_2021/Tract_Geometries
+	echo "Tracts Unzipped"
+else
+	echo "Tracts Unzipped"
+fi
 
-curl -o Data/NHTS_2017/csv.zip https://nhts.ornl.gov/assets/2016/download/csv.zip
+if [ ! -f Data/NHTS_2017/csv.zip ]; then
+	curl -o Data/NHTS_2017/csv.zip https://nhts.ornl.gov/assets/2016/download/csv.zip
+	echo "NHTS Data Downloaded"
+else
+	echo "NHTS Data Downloaded"
+fi
 
-unzip Data/NHTS_2017/csv.zip -d Data/NHTS_2017
+if [ ! -f Data/NHTS_2017/trippub.csv ]; then
+	unzip Data/NHTS_2017/csv.zip -d Data/NHTS_2017
+	echo "NHTS Data Unzipped"
+else
+	echo "NHTS Data Unzipped"
+fi
